Fix bug #3 bug clipping extreme badness instead of reject it.
An error in the algorithm was avoiding updating the best break-point candidate when exceeded 1.0, it should have been clipping instead so as to allow tracking of white-space which would otherwise be ignored. The fallback break point candidate was intended to perform this task but wasn't being used reliably and was essentially a hack to try to work around the actual bug.